Moved .hpp and updated .cpp
Moved .hpp files to include directory and updated .cpp files to reflect this. Deleted some temp txt files too.
