--- include/temp_file.txt
@@ -1,1 +1,0 @@
-temp file for git to tracking directory structure
