+++ src/action.cpp
@@ -1,4 +1,4 @@
-#include "action.hpp"
+#include "../include/action.hpp"
 
 void Action::run_action() {
     command->execute();
