+++ src/command.cpp
@@ -1,4 +1,4 @@
-#include "command.hpp"
+#include "../include/command.hpp"
 
 //simulates moving by changing the current room to the next room.
 void LeaveCommand::execute() {
